check_tracking/pyproject.toml
[project]
name = "check-tracking"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "pillow",
    "opencv-python",
    "opencv-contrib-python",
    "torch",
    "torchvision",
    "facenet-pytorch",
    "sixdrepnet",
    "ultralytics",
    "face_detection",
     "lap",             # used by ultralitics
    # "deepface",       # uses tensorflow
    "scipy-stubs~=1.17.1",
]



[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

